+++ tests/Filter/ScriptMinifierTest.php
@@ -1,0 +1,48 @@
+<?php
+
+namespace Staticka\Filter;
+
+/**
+ * Script Minifier Test
+ *
+ * @package Staticka
+ * @author  Rougin Royce Gutib <rougingutib@gmail.com>
+ */
+class ScriptMinifierTest extends \PHPUnit_Framework_TestCase
+{
+    /**
+     * @var string
+     */
+    protected $code;
+
+    /**
+     * @var \Staticka\Filter\FilterInterface
+     */
+    protected $filter;
+
+    /**
+     * Sets up the filter instance.
+     */
+    public function setUp()
+    {
+        $name = (string) str_replace('Filter', 'Fixture', __DIR__);
+
+        $this->filter = new ScriptMinifier;
+
+        $this->code = file_get_contents($name . '/Script.js');
+    }
+
+    /**
+     * Tests FilterInterface::filter.
+     *
+     * @return void
+     */
+    public function testFilterMethod()
+    {
+        $expected = 'map=function(){element=document.getElementById(\'rs-map\');place={lat:7.074744,lng:125.619585};map=new google.maps.Map(element,{zoom:17,center:place});marker=new google.maps.Marker({position:place,map:map});};';
+
+        $result = $this->filter->filter($this->code);
+
+        $this->assertEquals($expected, $result);
+    }
+}