+++ tests/Fixture/Layout.php
@@ -3,6 +3,11 @@
 <head>
   <meta charset="UTF-8">
   <title><?php echo $title; ?></title>
+  <style>
+    body {
+      background: #fff;
+    }
+  </style>
 </head>
 <body>
   <a href="<?php echo $url->set('home') ?>">Home</a>