+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 All notable changes to `Staticka` will be documented in this file.
 
-## [0.2.0](https://github.com/staticka/staticka/compare/v0.1.0...v0.2.0) - Unreleased
+## [0.2.0](https://github.com/staticka/staticka/compare/v0.1.0...v0.2.0) - 2018-04-07
+
+### Added
+- `Filter\ScriptMinifier`
+- `Filter\StyleMinifier`
+- `Helper\ViewHelper`
 
 ### Fixed
 - Error in transferring files with directories